fastapi-server/requirements.txt
fastapi
uvicorn
asyncpg
pandas
apscheduler
pyampd
python-dotenv
python-multipart
python-decouple
pydantic
pytest
cryptography
aiomysql
PyMySQL
pytz
ta
httpx
loguru
requests
beautifulsoup4
google-generativeai
selenium
webdriver-manager
numpy

pytest-asyncio
python-dateutil
aiofiles
aiohttp